For Go 1.12 and newer convertMap uses MapRange
