+++ packages/encoding.json.go
@@ -1,8 +1,8 @@
 package packages
 
 import (
-	"reflect"
 	"encoding/json"
+	"reflect"
 
 	"github.com/mattn/anko/env"
 )
