+++ vm/example_functions_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/mattn/anko/vm"
 	"github.com/mattn/anko/env"
+	"github.com/mattn/anko/vm"
 )
 
 func Example_vmFunctions() {
