+++ vm/main_test.go
