+++ vm/packages_test.go
